pdfdownloader/requirements.txt
playwright==1.40.0 
google-generativeai>=1.0.0 